Added simple demonstrations in "scripts/tutorials/04_sensors"

+++ source/isaaclab/isaaclab/sensors/ray_caster/ray_caster.py
@@ -229,7 +229,7 @@ class RayCaster(SensorBase):
         self._data.quat_w = torch.zeros(self._view.count, 4, device=self._device)
         self._data.ray_hits_w = torch.zeros(self._view.count, self.num_rays, 3, device=self._device)
         if self.cfg.track_ray_distance:
-            self._data.ray_distance = torch.full((self._view.count, self.num_rays), self.cfg.max_distance, device=self._device)
+            self._data.ray_distance = torch.full((self._view.count, self.num_rays), self.cfg.max_distance, device=self._device, dtype=torch.float32)
 
     def _update_buffers_impl(self, env_ids: Sequence[int]):
         """Fills the buffers of the sensor data."""
